backend/requirements.txt
Flask
pandas
python-dotenv
google-generativeai
Flask-Cors
pytest